Event listener
An active worker that listens continuously to the queue

RabbitListener annotation tells spring that "This method should be automatically invoked whenever a message appears in the transactions.queue" in other words "It's the "mail clerk" that Spring hires to constantly watch the transactions.queue P.O. Box and process any mail that arrives."

Post this, once a message is sent to the exchange with a routing key, the exchange sends a copy of that message to the queue linked/binded with that specific routing key, and this listener(aka worker) listens, processes and acknowledges the queue.

If the method completes successfully, automatically send an acknowledgement (ack) back to RabbitMQ to safely remove the message from the queue.

Exchange is only responsible for routing.